feat: Updated Docker-start.sh logic & made reusable components
